server/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dotenv
google-genai
